Remove deprecated servlet API calls
This is a hotfix, to enable our projects depending on Jersey to be migrated to Tomcat 10.1 using the Tomcat Jakarta Migration Tool
Tomcat 10.1 switched to Servlet API 6.0 which removed some deprecated methods. HttpServletRequest.setStatus(int,String) was used in the 2.x branch.

This commit is based on commit id 4977fec